# api/openapi.yaml
openapi: 3.0.0
info:
  title: Content Sharing Platform API
  description: |
    A comprehensive REST API for a content sharing platform that allows users to share, discover, and rate media content.
    
    ## Features
    - JWT-based authentication
    - User registration and login
    - Media content management (CRUD)
    - Rating system (1-5 stars)
    - Advanced filtering and search
    - Pagination support
    - Rating statistics
    
    ## Authentication
    Most endpoints require authentication using JWT tokens. After logging in or registering, include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
...
paths:
...
  /api/ratings:
    get:
      tags:
        - Ratings
      summary: List ratings
      description: Get a paginated list of ratings with optional filtering
      operationId: listRatings
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          description: Items per page
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: media_id
          in: query
          description: Filter by media content ID
          schema:
            type: integer
        - name: user_id
          in: query
          description: Filter by user ID
          schema:
            type: integer
      responses:
        '200':
          description: List of ratings
          content:
            application/json:
              schema:
                type: object
                properties:
                  ratings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Rating'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Ratings
      summary: Create rating
      description: Create a new rating for a media content (requires authentication)
      operationId: createRating
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingCreate'
      responses:
        '201':
          description: Rating created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Rating created successfully
                  rating:
                    $ref: '#/components/schemas/Rating'
        '400':
          description: Validation error or duplicate rating
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Media content not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/ratings/{rating_id}:
    get:
      tags:
        - Ratings
      summary: Get rating by ID
      description: Get detailed information about a specific rating
      operationId: getRating
      parameters:
        - name: rating_id
          in: path
          required: true
          description: Rating ID
          schema:
            type: integer
      responses:
        '200':
          description: Rating details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rating'
        '404':
          description: Rating not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Ratings
      summary: Update rating
      description: Update an existing rating (only by creator)
      operationId: updateRating
      security:
        - BearerAuth: []
      parameters:
        - name: rating_id
          in: path
          required: true
          description: Rating ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingUpdate'
      responses:
        '200':
          description: Rating updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Rating updated successfully
                  rating:
                    $ref: '#/components/schemas/Rating'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not the rating creator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Rating not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Ratings
      summary: Delete rating
      description: Delete a rating (only by creator)
      operationId: deleteRating
      security:
        - BearerAuth: []
      parameters:
        - name: rating_id
          in: path
          required: true
          description: Rating ID
          schema:
            type: integer
      responses:
        '200':
          description: Rating deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Rating deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not the rating creator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Rating not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/ratings/media/{media_id}/stats:
    get:
      tags:
        - Ratings
      summary: Get rating statistics
      description: Get comprehensive rating statistics for a media content
      operationId: getRatingStats
      parameters:
        - name: media_id
          in: path
          required: true
          description: Media content ID
          schema:
            type: integer
      responses:
        '200':
          description: Rating statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RatingStats'
        '404':
          description: Media content not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
...
    User:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
        username:
          type: string
          example: johndoe
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    RatingCreate:
      type: object
      required:
        - media_id
        - score
      properties:
        media_id:
          type: integer
          example: 1
          description: ID of the media content to rate
        score:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
          description: Rating score (1-5 stars)
        comment:
          type: string
          example: Excellent content!
          description: Optional comment about the rating
    RatingUpdate:
      type: object
      properties:
        score:
          type: integer
          minimum: 1
          maximum: 5
          example: 4
          description: Updated rating score (1-5 stars)
        comment:
          type: string
          example: Very good content
          description: Updated comment
    Rating:
      type: object
      properties:
        rating_id:
          type: integer
          example: 1
        media_id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        score:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: Excellent content!
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T15:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T15:00:00Z"
        user:
          $ref: '#/components/schemas/User'
        media:
          type: object
          properties:
            media_id:
              type: integer
              example: 1
            title:
              type: string
              example: Amazing Sunset Video
    RatingStats:
      type: object
      properties:
        media_id:
          type: integer
          example: 1
        total_ratings:
          type: integer
          example: 42
          description: Total number of ratings
        average_rating:
          type: number
          format: float
          example: 4.5
          description: Average rating score
        rating_distribution:
          type: object
          properties:
            '1':
              type: integer
              example: 2
              description: Number of 1-star ratings
            '2':
              type: integer
              example: 3
              description: Number of 2-star ratings
            '3':
              type: integer
              example: 5
              description: Number of 3-star ratings
            '4':
              type: integer
              example: 12
              description: Number of 4-star ratings
            '5':
              type: integer
              example: 20
              description: Number of 5-star ratings
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
          description: Current page number
        per_page:
          type: integer
          example: 10
          description: Items per page
        total_items:
          type: integer
          example: 42
          description: Total number of items
        total_pages:
          type: integer
          example: 5
          description: Total number of pages
    Error:
      type: object
      properties:
        error:
          type: string
          example: Validation error
          description: Error message
        message:
          type: string
          example: Invalid input data
          description: Detailed error description